backend/requirements.txt
# ── Core framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
# For PostgreSQL: uncomment the line below
# asyncpg>=0.29.0

# ── Settings ───────────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Security ───────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ── HTTP client (REST connector) ───────────────────────────────────────────────
httpx>=0.27.0

# ── Optional connectors ────────────────────────────────────────────────────────
# asyncio-mqtt>=0.16.0        # MQTT connector
# asyncua>=1.0.0              # OPC-UA connector

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
anyio>=4.3.0

# ── Logging / tracing ─────────────────────────────────────────────────────────
python-multipart>=0.0.9
